samples/durable-task-sdks/python/arXiv_research_agent/requirements.txt
# arXiv Research Agent - Python Dependencies

# DurableTask SDK for Azure
durabletask-azuremanaged>=1.3.0

# OpenAI SDK (Responses API)
openai>=1.0.0

# Azure Identity for Entra ID authentication (optional)
azure-identity>=1.15.0

# FastAPI for REST API
fastapi>=0.109.0
uvicorn>=0.27.0

# HTTP client
httpx>=0.26.0

# Environment variables
python-dotenv>=1.0.0

# Rich console output (optional, for CLI)
rich>=13.0.0

# Pydantic for data validation (included with FastAPI)
pydantic>=2.5.0

# Testing dependencies
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0
